requirements.txt
alphabet-detector==0.0.7
appnope==0.1.3
attrs==23.1.0
awscrt==0.19.19
Babel==2.9.1
beautifulsoup4==4.12.2
bleach==6.1.0
cachetools==5.3.2
<<<<<<< HEAD
certifi>=2024.07.04
cffi==1.14.5
=======
certifi==2023.7.22
cffi==1.16.0
>>>>>>> 0c3df24 (updated requirements and gpdb-cleanup.py to utilize nvidia gpu)
chardet==3.0.4
charset-normalizer==3.3.2
chevron==0.13.1
colorama==0.4.4
configobj==5.0.6
cryptography==36.0.1
decorator==5.1.1
defusedxml==0.7.1
distro==1.5.0
docutils==0.16
entrypoints==0.4
fastjsonschema==2.18.1
ffmpeg-python==0.2.0
filelock==3.13.1
fsspec==2023.12.2
future==0.18.3
google-api-core==2.15.0
google-auth==2.26.2
google-cloud-core==2.4.1
google-cloud-translate==3.14.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
idna==2.8
ipython-genutils==0.2.0
Jinja2==3.1.2
jmespath==0.10.0
jsonpatch==1.21
jsonpointer==2.0
jsonschema==3.2.0
jupyter_core==5.4.0
jupyterlab-pygments==0.2.2
llvmlite==0.41.1
lockfile==0.12.2
MarkupSafe==2.1.3
mistune==3.0.2
more-itertools==10.2.0
mpmath==1.3.0
nest-asyncio==1.5.8
netifaces==0.10.6
networkx==3.2.1
numba==0.58.1
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.0.2
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.0
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==3.11.0
ply==3.11
portpicker==1.2.0
prettytable==0.7.2
prompt-toolkit==1.0.18
proto-plus==1.23.0
protobuf==4.25.2
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.20
Pygments==2.16.1
pyrsistent==0.17.3
pyserial==3.4
PySocks==1.7.1
python-daemon==2.3.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
regex==2023.12.25
requests==2.31.0
rpds-py==0.10.6
rsa==4.9
simplegeneric==0.8.1
six==1.12.0
soupsieve==2.5
stopit==1.1.2
sympy==1.12
terminado==0.13.3
textdistance==4.6.1
tiktoken==0.5.2
tinycss2==1.2.1
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
tornado==4.5.3
tqdm==4.66.1
traitlets==5.12.0
triton==2.2.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.24.3
wcwidth==0.2.8
webencodings==0.5.1
